Move selective regularization to sklearn_extensions
